open_ai: Move `responses` module to its own file
